Update README.md
The healthcare sector is witnessing a rapid digital transformation, emphasizing enhancing
patient experience and operational efficiency. One key area where technology can make a
significant impact is the process of booking appointments with healthcare providers.
Traditional methods often result in long waiting times and inefficiencies in scheduling,
highlighting the need for a more streamlined approach.The primary objective of the system is to provide a user-friendly interface for patients to
search for nearby doctors, view their profiles and availability, and book appointments
seamlessly. For healthcare providers, the system aims to simplify the management of
appointments, profiles, and communication with patients. Additionally, the system will
include administrative features for managing user accounts and appointments.